Add connection-owned cells
Cells can be marked by connection id and will be removed from notebooks
as soon as the connection drops.

Such cells won't be persisted to a filesystem obviously.

The main use case is to attach reactive computations to notebooks which
are used for ephemeral UI state.

Right now such cells are being sent to all connected clients but a
further future improvement could be to send such cells only to relevant
clients.